cambiando a una matriz de 3x3, lista el determinante, tambien cambió la estructura de los datos, ya no usé arreglos :p

+++ decorator/matriz/Matriz.java
@@ -1,4 +1,4 @@
 public interface Matriz {
   public void mostrar_valores();
-  public int determinante();
+  public double determinante();
 }
